+++ src/apps/chifra/pkg/monitor/filter.go
@@ -62,6 +62,7 @@ func (mon *Monitor) ReadAndFilterAppearances(filt *filter.AppearanceFilter) (app
 	apps = make([]types.SimpleAppearance, 0, len(fromDisc))
 	for _, app := range fromDisc {
 		app := app
+		// #WITHDRAWALS
 		if app.TransactionId == uint32(types.Withdrawal) {
 			continue
 		}
@@ -75,8 +76,8 @@ func (mon *Monitor) ReadAndFilterAppearances(filt *filter.AppearanceFilter) (app
 			filt.OuterBounds.Last = uint64(app.BlockNumber + 1)
 			s := types.SimpleAppearance{
 				Address:          mon.Address,
-				BlockNumber:      uint32(app.BlockNumber),
-				TransactionIndex: uint32(app.TransactionId),
+				BlockNumber:      app.BlockNumber,
+				TransactionIndex: app.TransactionId,
 				Timestamp:        utils.NOPOSI,
 			}
 			apps = append(apps, s)
